Changing the window object to window.matchMedia
