+++ src/hooks/index.js
@@ -1,1 +1,1 @@
-export { useIsMobile } from "./use-is-mobile";
+export { useMediaQuery } from "./use-media-query";
